[front/components] - refactor: streamline ManageUsersPanel state management
 - Removed useRef hook in favor of useState for managing user selections
 - Simplified logic related to editor-only mode by removing unnecessary map for tracking users
 - Updated MemberSelectionTable component to handle user selection changes internally

[front/components] - fix: ensure canSave logic in ManageUsersPanel is concise and correct

 - Fixed the canSave condition to properly account for both "space-members" and "editors-only" modes

[front/components] - refactor: update MemberSelectionTable to initialize user map with props

 - Refactored MemberSelectionTable to accept initial members as a prop instead of handling them in the parent component
 - Removed onMembersLoaded callback and incorporated logic into useEffect for maintaining the internal user map

[front/components] - refactor: clean up RestrictedAccessBody by removing outdated code

 - Removed unnecessary user map reference and related logic simplifying member selection changes handling
